ci: address nullaway review feedback

+++ gradle-plugin/src/main/java/media/barney/cognitive/gradle/CognitiveJavaCheckTask.java
@@ -756,7 +756,7 @@ public abstract class CognitiveJavaCheckTask extends DefaultTask {
         if (isCurrentRememberedPath(rememberedReport, reportPath)) {
             return;
         }
-        if (reportChanged(reportPath, before)) {
+        if (reportPath != null && reportChanged(reportPath, before)) {
             Files.deleteIfExists(reportPath);
         }
     }
